fix: Change name scheme for service name.
We follow the principle not using configurable parameters in the service name. Instead provisioning a service like
HTTP-80 we provision a new by changing the port which we don't want. We use instead a service name like HTTP.MyApp
where we can change configuration parameters on the same service name.